Bug#32679024 PURGE CAN REMOVE TOO MUCH BECAUSE OF MISSING FENCE
The view->m_closed was not an atomic variable, yet it was modified from
user transaction thread (when considering re-opening the view)
concurrently with the purge thread reading it (to find the oldest open
read-view).
In theory:
* this is an UB
* setting it to false and proceeding to next operation without memory
  barrier, means the CPU can have the write still in the store buffer
  which defeats the goal of informing the purge thread we might reopen
  the view
In practice:
* we never seen any crash due to this particular race-condition
* there is a complicated argument showing that at least on x86/64 it is
  near impossible for the whole scenario to fit in time it takes to
  execute just few lines (as context switch flushes store buffer).

This patch, changes the view->m_closed to atomic to avoid race
conditions, and make the reasoning much simpler.

Change-Id: Idb8503d6c0297cec01c15ae0ec0e2f749151d5c6